ensure custom culture invariant for TypeAnnotatedObjectConverter
