+++ Project-Aurora/Project-Aurora/Utils/Json/JSONUtils.cs
@@ -263,110 +263,6 @@ public class OverrideTypeConverter : JsonConverter
     }
 }
 
-public class TypeAnnotatedObjectConverter : JsonConverter
-{
-    public override bool CanConvert(Type objectType) => objectType.FullName == typeof(Color).FullName ||
-                                                        objectType.IsAssignableFrom(typeof(object));
-
-    public override void WriteJson(JsonWriter writer, object? value, JsonSerializer serializer)
-    {
-        if (value == null)
-        {
-            writer.WriteNull();
-            return;
-        }
-        writer.WriteStartObject();
-        writer.WritePropertyName("$type");
-        writer.WriteValue(value.GetType().AssemblyQualifiedName);
-        writer.WritePropertyName("$value");
-        serializer.Serialize(writer, value);
-        writer.WriteEndObject();
-    }
-
-    public override object? ReadJson(JsonReader reader, Type objectType, object? existingValue, JsonSerializer serializer)
-    {
-        // If this is a null token, then the original value was null.
-        var readerTokenType = reader.TokenType;
-        if (readerTokenType == JsonToken.StartObject)
-        {
-            return ParseJsonNode(reader, objectType, serializer);
-        }
-        
-        var json = reader.Value?.ToString();
-        if (string.IsNullOrWhiteSpace(json))
-        {
-            return null;
-        }
-        switch (readerTokenType)
-        {
-            case JsonToken.String:
-                return json.StartsWith('\"')
-                    ? JsonConvert.DeserializeObject(json, objectType)
-                    : JsonConvert.DeserializeObject("\"" + json + "\"", objectType);
-            case JsonToken.Integer:
-                return JsonConvert.DeserializeObject(json, objectType);
-            case JsonToken.Boolean:
-                switch (json.ToLowerInvariant())
-                {
-                    case "true":
-                        return true;
-                    case "false":
-                        return false;
-                    default:
-                        return false;
-                }
-            case JsonToken.Null:
-                return existingValue;
-        }
-
-        Global.logger.Warning("Unhandled json type {Token}", readerTokenType);
-        return JsonConvert.DeserializeObject(json, objectType);
-    }
-
-    private static object? ParseJsonNode(JsonReader reader, Type objectType, JsonSerializer serializer)
-    {
-        var item = serializer.Deserialize<JObject>(reader);
-        var jToken = item?["$type"];
-        if (item == null || jToken == null)
-        {
-            throw new JsonReaderException("item or $type is null");
-        }
-        var type = serializer.Deserialize<Type>(jToken.CreateReader());
-        var value = item["$value"];
-        if (value == null)
-        {
-            return serializer.Deserialize(item.CreateReader(), type);
-        }
-
-        var valueReader = value.CreateReader();
-        switch (valueReader.TokenType)
-        {
-            case JsonToken.StartObject:
-                return serializer.Deserialize(valueReader, type);
-            default:
-                var s = value.ToString();
-                if (type == typeof(bool) || type == typeof(Color))
-                {
-                    var colorString = value.ToString();
-                    if (colorString.StartsWith('"'))
-                    {
-                        return JsonConvert.DeserializeObject(value.ToString(), type);
-                    }
-                    return JsonConvert.DeserializeObject("\"" + value + "\"", type);
-                }
-                if (objectType.FullName != typeof(Color).FullName && type?.FullName != typeof(Color).FullName)
-                    return JsonConvert.DeserializeObject(s, type);
-                if (s.StartsWith('\"'))
-                {
-                    return JsonConvert.DeserializeObject(s, type);
-                }
-
-                Global.logger.Error("Attempting to convert unknown type: {Type}", type);
-                return JsonConvert.DeserializeObject("\"" + value + "\"", type);
-        }
-    }
-}
-
 public class SingleToDoubleConverter : JsonConverter
 {
     public override bool CanConvert(Type objectType) => objectType == typeof(double);
